Add python support

+++ src/Build.ts
@@ -168,12 +168,16 @@ export default class Build {
         {
             console.log(`building ${name} in ${buildPath}`);
 
-            // TODO: merge dependencies from template
-            const execResult = await this.execPromise("pip install -r requirements.txt --target .", 
-                { cwd: buildPath, env: process.env });
+            if( fsUtils.exists(path.join(buildPath, 'requirements.txt')) ) {
+                console.log('installing dependencies...')
+                const execResult = await this.execPromise("pip install -r requirements.txt -t .", 
+                    { cwd: buildPath, env: process.env });
 
-            if (execResult.stderr) {
-                throw new Error(`pip install returned an error:\n${execResult.stdout}\n${execResult.stderr}`);
+                if (execResult.stderr) {
+                    throw new Error(`pip install returned an error:\n${execResult.stdout}\n${execResult.stderr}`);
+                }
+            } else {
+                console.log('no requirements.txt file. skipping pip install.')
             }
             
         } catch (err) {
